+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [0.4.0] - 2026-01-14
 - fix #6, typo in UVB and UVC wrapper functions.
+- add **uint16_t getRawUVA(), getRawUVB(), getRawUVC()** - see #6
+- add example **getRawUV()**
 - update GitHub actions
 - minor edits